# docs/openapi.yaml
openapi: 3.0.3
info:
  title: e-Food Center API
  version: 0.1.0
  description: |
    MVP REST API contract for e-Food Center.
    Source of truth for parallel mobile, admin, and backend work.
    Normalized envelope: `{ success, data, error, meta }`.
  contact:
    name: Somnath Das
servers:
  - url: http://localhost:4000/api/v1
    description: Local Dev
  - url: https://staging-api.example.com/api/v1
    description: Staging
  - url: https://api.example.com/api/v1
    description: Production
...
paths:
...
  /auth/otp/request:
    post:
      tags: [Auth]
      summary: Request phone OTP
      operationId: requestOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  example: '919876543210'
                  description: E.164 without + or with country code
      responses:
        '200':
          description: OTP sent (always generic success to avoid phone enumeration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP and issue tokens
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, otp]
              properties:
                phone: { type: string }
                otp: { type: string, minLength: 4, maxLength: 8 }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New tokens
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /categories:
    get:
      tags: [Catalog]
      summary: List categories
      operationId: listCategories
      parameters:
        - $ref: '#/components/parameters/branchId'
      responses:
        '200':
          description: Categories
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'
  /products:
    get:
      tags: [Catalog]
      summary: List products (guest OK)
      operationId: listProducts
      parameters:
        - $ref: '#/components/parameters/branchId'
        - name: categoryId
          in: query
          schema: { type: string, format: uuid }
        - name: q
          in: query
          description: Search query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Product page
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
                      meta:
                        $ref: '#/components/schemas/PageMeta'
  /products/{productId}:
    get:
      tags: [Catalog]
      summary: Product detail
      operationId: getProduct
      parameters:
        - $ref: '#/components/parameters/productId'
        - $ref: '#/components/parameters/branchId'
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    branchId:
      name: branchId
      in: query
      schema: { type: string, format: uuid }
...
    productId:
      name: productId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiSuccess:
      type: object
      required: [success, data, error, meta]
      properties:
        success: { type: boolean, example: true }
        data: {}
        error: { nullable: true }
        meta:
          type: object
          additionalProperties: true
    ApiError:
      type: object
      required: [success, data, error, meta]
      properties:
        success: { type: boolean, example: false }
        data: { nullable: true }
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
        meta:
          type: object
          additionalProperties: true
    PageMeta:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        requestId: { type: string }
    AuthTokens:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, description: Access token TTL seconds }
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        phone: { type: string }
        email: { type: string, nullable: true }
        role:
          type: string
          enum: [customer, staff, manager, admin]
        preferredLanguage:
          type: string
          enum: [en, bn]
...
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        sortOrder: { type: integer }
    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        sku: { type: string }
        name: { type: string }
        description: { type: string }
        categoryId: { type: string, format: uuid }
        price: { type: number }
        minQty: { type: integer, minimum: 1 }
        imageUrl: { type: string, nullable: true }
        isAvailable: { type: boolean }
        availableQty: { type: integer, nullable: true }
...